% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_het_components_fit.R
\name{hom_het_model}
\alias{hom_het_model}
\title{Fitting shared and specific components for each tissue or cell type}
\usage{
hom_het_model(
  X = NULL,
  Ys = NULL,
  covs = NULL,
  X_file = NULL,
  Y_file_dir = NULL,
  cov_file_dir = NULL,
  verbose = T,
  out_dir = NULL,
  snps = NULL,
  signif = 0.1,
  gene_name = NULL,
  content_alpha = 0.5,
  tissue_alpha = 0.5,
  num_folds = 10,
  scale_exp = F,
  seed = 9000
)
}
\arguments{
\item{X}{A genotype matrix, individuals are row names, no column names.}

\item{Ys}{Expression data matrix, individuals are row names, no column names.}

\item{covs}{Covariates data.}

\item{X_file}{Direction containing genotype matrix, individuals are row names, no column names.}

\item{Y_file_dir}{Direction containing only the expression files, individuals are row names, no column names.}

\item{cov_file_dir}{Specify this directory if you want to residualize the expression. Should share the same prefix as the expression files}

\item{verbose}{Provide more information.}

\item{out_dir}{Output CV predictors and other statistics.}

\item{snps}{A tab-delimited .txt file containing snp information. six columns.}

\item{signif}{Nominally significant p-value threshold (alpha). default=0.1}

\item{gene_name}{Add this prefix to all the saved results files. default=NULL}

\item{content_alpha}{Regularization constant, 1 means LASSO regression.}

\item{tissue_alpha}{Regularization constant for the tissue by tissue apprach.}

\item{num_folds}{Number of folds for cross-validation.}

\item{scale_exp}{Center and scale the gene expression}

\item{seed}{Seed for analysis}
}
\value{
This function returns the estimated shared and specific components and other coefficients
}
\description{
Fitting shared and specific components for each tissue or cell type.
}
