% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_fit_using_hom_het_components.R
\name{Model_fit}
\alias{Model_fit}
\title{Fitting models using shared and specific components generated in the first step}
\usage{
Model_fit(Gname)
}
\arguments{
\item{Gname}{Add this prefix to all the saved results files.}
}
\value{
This function returns the estimated coefficients in the second step
}
\description{
Fitting models using shared and specific components generated in the first step.
}
