% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_weights.R
\name{Weights_gen}
\alias{Weights_gen}
\title{Generating weights files required for FUSION TWAS}
\usage{
Weights_gen(G_name, in_dir, out_dir)
}
\arguments{
\item{G_name}{Add this prefix to all the saved results files.}

\item{in_dir}{Direction for files stored from about two steps.}

\item{out_dir}{Direction for output weights files.}
}
\value{
This function generates weights files required for FUSION TWAS.
}
\description{
Generating weights files required for FUSION TWAS.
}
