% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights_Pre.R
\name{Weights_pre}
\alias{Weights_pre}
\title{Combing coefficients from above two steps and prepare for weights file}
\usage{
Weights_pre(G_name, in_dir)
}
\arguments{
\item{G_name}{Add this prefix to all the saved results files.}

\item{in_dir}{Direction for files stored from about two steps.}
}
\value{
This function returns the combined coefficients from above two steps.
}
\description{
Combing coefficients from above two steps and prepare for weights file.
}
